refactoring of point2 and point3 ... now they would be more feasible to run. Too much effort to change how the point4 run the thompson sampling algorithm, maybe better to leave it as it is.
Many tests on point2 and point3, only with 100-150 nodes (15 minutes for running the point 3 on 3 graphs with these dimensions...).

The point4 doesn't change its results if we change simulations, instead we have big improvements if we increase the repetitions!

I am not able to touch the point5... I dunno what it does!